--- src/components/myFavorite/MyFavorite.jsx
@@ -1,4 +1,0 @@
-const MyFavorite = () => {
-  return <div></div>;
-};
-export default MyFavorite;
